// backend/src/targeting-rules/targeting-rule-operator.enum.ts
export enum TargetingRuleOperator {
  Equals = "EQUALS",
  NotEquals = "NOT_EQUALS",
  Contains = "CONTAINS",
  NotContains = "NOT_CONTAINS",
  StartsWith = "STARTS_WITH",
  EndsWith = "ENDS_WITH",
  In = "IN",
  NotIn = "NOT_IN",
  GreaterThan = "GREATER_THAN",
  GreaterThanOrEqual = "GREATER_THAN_OR_EQUAL",
  LessThan = "LESS_THAN",
  LessThanOrEqual = "LESS_THAN_OR_EQUAL"
}